[1/x] Refactor default evaluator (#13360)
