+++ mlflow/models/evaluation/evaluator_registry.py
@@ -12,10 +12,16 @@ class ModelEvaluatorRegistry:
 
     def __init__(self):
         self._registry = {}
+        self._builtin_evaluators = {}
 
     def register(self, scheme, evaluator):
         """Register model evaluator provided by other packages"""
         self._registry[scheme] = evaluator
+
+    def register_builtin(self, scheme, evaluator):
+        """Register built-in model evaluator"""
+        self._registry[scheme] = evaluator
+        self._builtin_evaluators[scheme] = evaluator
 
     def register_entrypoints(self):
         # Register ModelEvaluator implementation provided by other packages
@@ -42,14 +48,31 @@ class ModelEvaluatorRegistry:
             )
         return evaluator_cls()
 
+    def is_builtin(self, name):
+        return name in self._builtin_evaluators
+
+    def is_registered(self, name):
+        return name in self._registry
+
 
 _model_evaluation_registry = ModelEvaluatorRegistry()
 
 
 def register_evaluators(module):
-    from mlflow.models.evaluation.default_evaluator import DefaultEvaluator
+    from mlflow.models.evaluation.evaluators.classifier import ClassifierEvaluator
+    from mlflow.models.evaluation.evaluators.default import DefaultEvaluator
+    from mlflow.models.evaluation.evaluators.regressor import RegressorEvaluator
+    from mlflow.models.evaluation.evaluators.shap import ShapEvaluator
 
-    module._model_evaluation_registry.register("default", DefaultEvaluator)
+    # Built-in evaluators
+    module._model_evaluation_registry.register_builtin(DefaultEvaluator.name, DefaultEvaluator)
+    module._model_evaluation_registry.register_builtin(
+        ClassifierEvaluator.name, ClassifierEvaluator
+    )
+    module._model_evaluation_registry.register_builtin(RegressorEvaluator.name, RegressorEvaluator)
+    module._model_evaluation_registry.register_builtin(ShapEvaluator.name, ShapEvaluator)
+
+    # Plugin evaluators
     module._model_evaluation_registry.register_entrypoints()
